move TestAgentEntryToJSON to entry_agent_test.go
